+++ src/GuildrunAccess.Module/GameRun/Hud/PartySection.cs
@@ -71,7 +71,7 @@ namespace GuildrunAccess.Module.GameRun
                     OnFocus = () => { if (RunData.TryHeroId(view, out var id)) HeroActions.PeekHero(id); },
                     Details = () => SlotDetails(view),
                     Subject = () => view,
-                    SideLines = HeroLines.Side(() => SlotRows(view), () => ItemNodes.ItemTooltips(view._itemSlotViews)),
+                    SideLines = HeroLines.SideOfSlots(() => SlotRows(view), () => view._itemSlotViews),
                 });
             }
             b.PopContext();